Used max warning level to fix some minor bugs and stylistic issues

+++ tests/permutation_tests.cpp
@@ -8,7 +8,7 @@ using namespace dscr;
 void check_permutation(permutations::permutation x)
 {
 	std::sort(x.begin(), x.end());
-	for (int i = 0; i < x.size(); ++i)
+	for (size_t i = 0; i < x.size(); ++i)
 	{
 		ASSERT_EQ(x[i],i);
 	}
